+++ src/meegsim/_check.py
@@ -462,7 +462,7 @@ def check_coupling_params(method, coupling_params, coupling_edge):
     """
 
     # Test on a 10 second segment of white noise
-    sfreq = 100
+    sfreq = 500
     rng = np.random.default_rng(seed=0)
     waveform = rng.random((10 * sfreq,))
 
@@ -598,6 +598,21 @@ def check_extents(extents, n_sources):
                 )
 
     return extents
+
+
+def check_stc_as_param(stc, src):
+    for src_idx, s in enumerate(src):
+        common = np.intersect1d(stc.vertices[src_idx], s["vertno"], assume_unique=True)
+
+        # XXX: the code below overlaps with sources._BaseSource.check_compatibility
+        missing_vertno = set(s["vertno"]) - set(common)
+        if missing_vertno:
+            report_missing = ", ".join([str(v) for v in missing_vertno])
+            raise ValueError(
+                f"The provided stc does not contain all vertices of the "
+                f"source space that is used for simulations. The following vertices "
+                f"from src[{src_idx}] are missing: {report_missing}"
+            )
 
 
 def check_colors(colors):