+++ src/meegsim/sources.py
@@ -8,7 +8,13 @@ to the original time series.
 import numpy as np
 import mne
 
-from .utils import vertices_to_mne, _extract_hemi, _hemi_to_index
+from meegsim.utils import (
+    vertices_to_mne,
+    _extract_hemi,
+    _get_center_of_mass,
+    _get_param_from_stc,
+    _hemi_to_index,
+)
 
 
 class _BaseSource:
@@ -18,12 +24,14 @@ class _BaseSource:
 
     kind = "base"
 
-    def __init__(self, waveform, std=1.0):
-        # Current constraint: one source corresponds to one waveform
-        # Point source: the waveform is present in one vertex
-        # Patch source: the waveform is mixed with noise in several vertices
+    def __init__(self, src_idx, waveform, std=1.0, name=""):
+        # Current constraints:
+        #  - one source corresponds to one waveform
+        #  - all vertices belong to the same source space (e.g., hemisphere)
+        self.src_idx = src_idx
         self.waveform = waveform
         self.std = std
+        self.name = name
 
     @property
     def data(self):
@@ -69,26 +77,60 @@ class _BaseSource:
                 f"contain the following vertices: {report_missing}"
             )
 
+    def to_label(self, src):
+        """
+        Get an mne.Label object containing all vertices belonging to the current
+        source.
+
+        Parameters
+        ----------
+        src : SourceSpaces
+            The source space where the source should be considered.
+
+        Returns
+        -------
+        label : Label
+            The constructed label.
+        """
+        self._check_compatibility(src)
+
+        # Make sure that we can turn the source into a label
+        if src[self.src_idx]["type"] != "surf":
+            raise ValueError(
+                "Only sources in surface source spaces can be converted into a label"
+            )
+
+        # NOTE: we need to sort vertices before constructing the label
+        vertno = np.atleast_1d(np.sort(self.vertices[:, 1]))
+        return mne.Label(
+            vertices=vertno,
+            pos=src[self.src_idx]["rr"][vertno, :],
+            # XXX: we should pass weights here once we implement patches with
+            # amplitude decay
+            values=np.ones_like(vertno),
+            hemi="rh" if self.src_idx else "lh",
+            name=self.name,
+        )
+
     def to_stc(self, src, tstep, subject=None):
         """
-        Convert the point source into a SourceEstimate object in the context
+        Convert the source into a SourceEstimate object in the context
         of the provided SourceSpaces.
 
         Parameters
         ----------
-        src: mne.SourceSpaces
-            The source space where the point source should be considered.
-        tstep: float
+        src : SourceSpaces
+            The source space where the source should be considered.
+        tstep : float
             The sampling interval of the source time series (1 / sfreq).
-        subject: str or None, optional
+        subject : str or None, optional
             Name of the subject that the stc corresponds to.
             If None, the subject name from the provided src is used if present.
 
         Returns
         -------
-        stc: mne.SourceEstimate
-            SourceEstimate that corresponds to the provided src and contains
-            one active vertex.
+        stc : SourceEstimate
+            SourceEstimate that corresponds to the source in the provided src.
 
         Raises
         ------
@@ -116,23 +158,25 @@ class PointSource(_BaseSource):
 
     Attributes
     ----------
-    src_idx: int
+    name : str
+        The name of source.
+    src_idx : int
         The index of source space that the point source belong to.
-    vertno: int
+    vertno : int
         The vertex that the point source correspond to
-    waveform: np.array
+    waveform : array
         The waveform of source activity.
-    hemi: str or None, optional
+    std : float, optional
+        The standard deviation of the source activity (1 by default).
+    hemi : str or None, optional
         Human-readable name of the hemisphere (e.g, lh or rh).
     """
 
     kind = "point"
 
     def __init__(self, name, src_idx, vertno, waveform, std=1.0, hemi=None):
-        super().__init__(waveform, std)
+        super().__init__(src_idx, waveform, std, name)
 
-        self.name = name
-        self.src_idx = src_idx
         self.vertno = vertno
         self.hemi = hemi
 
@@ -150,7 +194,7 @@ class PointSource(_BaseSource):
         return np.atleast_2d(np.array([self.src_idx, self.vertno]))
 
     @classmethod
-    def create(
+    def _create(
         cls, src, times, n_sources, location, waveform, stds, names, random_state=None
     ):
         """
@@ -174,6 +218,10 @@ class PointSource(_BaseSource):
             raise ValueError("The number of sources in waveform does not match")
         if data.shape[1] != len(times):
             raise ValueError("The number of samples in waveform does not match")
+
+        # Get the std values if an stc was provided
+        if isinstance(stds, mne.SourceEstimate):
+            stds = _get_param_from_stc(stds, vertices)
 
         # Create point sources and save them as a group
         sources = []
@@ -200,23 +248,25 @@ class PatchSource(_BaseSource):
 
     Attributes
     ----------
-    src_idx: int
+    name : str
+        The name of source.
+    src_idx : int
         The index of source space that the patch source belong to.
-    vertno: list
+    vertno : list
         The vertices that the patch sources correspond to including the central vertex.
-    waveform: np.array
+    waveform : np.array
         The waveform of source activity.
-    hemi: str or None, optional
+    std : float, optional
+        The standard deviation of the source activity (1 by default).
+    hemi : str or None, optional
         Human-readable name of the hemisphere (e.g, lh or rh).
     """
 
     kind = "patch"
 
     def __init__(self, name, src_idx, vertno, waveform, std=1.0, hemi=None):
-        super().__init__(waveform, std)
+        super().__init__(src_idx, waveform, std, name)
 
-        self.name = name
-        self.src_idx = src_idx
         self.vertno = vertno
         self.hemi = hemi
 
@@ -229,14 +279,17 @@ class PatchSource(_BaseSource):
 
     @property
     def data(self):
-        return np.tile(self.waveform, (len(self.vertno), 1))
+        # NOTE: the scaling factor is introduced to make the total variance of
+        # patch activity invariant to the number of vertices in the patch
+        scaling_factor = 1 / np.sqrt(len(self.vertno))
+        return np.tile(self.waveform, (len(self.vertno), 1)) * scaling_factor
 
     @property
     def vertices(self):
         return np.array([[self.src_idx, v] for v in self.vertno])
 
     @classmethod
-    def create(
+    def _create(
         cls,
         src,
         times,
@@ -273,13 +326,22 @@ class PatchSource(_BaseSource):
         # find patch vertices
         subject = src[0].get("subject_his_id", None)
         patch_vertices = []
+        patch_stds = [] if isinstance(stds, mne.SourceEstimate) else stds
         for isource, extent in enumerate(extents):
             src_idx, vertno = vertices[isource]
+            vertno = vertno if isinstance(vertno, list) else [vertno]
+
+            # Get the std values if an stc was provided
+            # NOTE: for now, we always use the value that corresponds to the
+            # center of the patch but we could allow the user to select another
+            # vertex
+            if isinstance(stds, mne.SourceEstimate):
+                center_vertno = _get_center_of_mass(src, src_idx, vertno)
+                std = _get_param_from_stc(stds, [(src_idx, center_vertno)])
+                patch_stds.append(std)
 
             # Add vertices as they are if no extent provided
             if extent is None:
-                # Wrap vertno in a list if it is a single number
-                vertno = vertno if isinstance(vertno, list) else [vertno]
                 patch_vertices.append(vertno)
                 continue
 
@@ -297,7 +359,7 @@ class PatchSource(_BaseSource):
         # Create patch sources and save them as a group
         sources = []
         for (src_idx, _), patch_vertno, waveform, std, name in zip(
-            vertices, patch_vertices, data, stds, names
+            vertices, patch_vertices, data, patch_stds, names
         ):
             hemi = _extract_hemi(src[src_idx])
             sources.append(